requirements.txt
streamlit
torch
transformers
accelerate
python-dotenv
openpyxl
librosa
soundfile
sentencepiece
protobuf
bitsandbytes
